更新角色时增加角色对应菜单列表的同步修改，权限表关联为user-sysrole-sysmenu 工会管理-机构管理： 新增机构各类操作接口 新增机构批量导入及模板下载接口 新增机构角色批量导入及模板下载接口 工会管理-会员管理： 新增用户各类操作接口

+++ ruoyi/src/main/java/com/ruoyi/project/union/entity/User.java
@@ -149,7 +149,7 @@ public class User extends Model<User> {
      * 帐号锁定期限
      */
     @TableField("lockDeadline")
-    private Integer lockDeadline;
+    private Long lockDeadline;
 
     /**
      * 帐号密码错误次数
